From Ground Zero.
First as we assemble, we attempt to make an instance of the SUT which is SignupFormModelValidator but since it does not exist - the test will fail.

Second go and create the basic Class structure for SignupFormModelValidator to allow us to assemble.

Third Act on the SignupFormModelValidator with our first check, which is isFirstNameValid - since this method does not exist the test will fail.

Go and create the function. With a least one reason to not validate the firstname, if it is empty.

Assert that when we provide a non empty string - it will return true.